requirements.txt
fastapi>=0.111.0
uvicorn[standard]>=0.29.0
python-multipart>=0.0.9
httpx>=0.27.0
motor>=3.4.0
pymongo>=4.7.0
python-dotenv>=1.0.1